Moved do_not_pickle_attributes inot utils
